% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGuide.R
\name{checkGuide}
\alias{checkGuide}
\title{Check a guide}
\usage{
checkGuide(guideId, url)
}
\arguments{
\item{guideId}{Id of your guide}

\item{url}{URL of your webpage}
}
\value{
Json with

orange, red, blue and green contain for each term to optimize two values: the field start value and the field end value.
scores contains the location of each term on the graph defined by the previous fields.

score : 100% for a text-optimized to the maximum of the normal range
danger : SEO risk score
}
\description{
Check a guide
}
\details{
Rescode
401 : invalid key or ip
429 : Too Many Requests
}
\examples{
\dontrun{
scores <- checkGuide(28183,"http://www.france24.fr")
}
}
\author{
Vincent Terrasi
}
