% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStatus.R
\name{checkStatus}
\alias{checkStatus}
\title{Sends back an empty ok json}
\usage{
checkStatus()
}
\value{
a json with ok, the current system load and timestamp
}
\description{
Sends back an empty ok json
}
\details{
ResCode
401 : Unauthorized : user or password does not match.
406 : Request and response must be application/json
}
\examples{
\dontrun{
status <- checkStatus()
}
}
\author{
Vincent Terrasi
}
