% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{countExpression}
\alias{countExpression}
\title{count each expression in a text}
\usage{
countExpression(text, searchword)
}
\value{
Number of expression
}
\description{
count each expression in a text
}
\details{

}
\examples{
\dontrun{
df <- countExpression(text,searchword)
}
}
\author{
Vincent Terrasi
}
