% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createToken.R
\name{createToken}
\alias{createToken}
\title{Requests a token for API usage}
\usage{
createToken()
}
\value{
a Token, valid for 2 hours
}
\description{
Requests a token for API usage
}
\details{
ResCode
401 : Unauthorized : user or password does not match.
406 : Request and response must be application/json
}
\examples{
\dontrun{
createToken()
}
}
\author{
Vincent Terrasi
}
