% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extractText}
\alias{extractText}
\title{Util Function for extracting text from URL}
\usage{
extractText(url, minchar, filter = FALSE)
}
\arguments{
\item{url}{URL}

\item{minchar}{Minimal length for a sentence}

\item{filter}{Boolean for basic or strict filtering}
}
\value{
A character with the main text
}
\description{
Util Function for extracting text from URL
}
\details{

}
\examples{
\dontrun{
extractText("https//www.wikipedia.com",2O,filter=FALSE)
extractText("https//www.wikipedia.com",2O,filter=TRUE)
}
}
\author{
Vincent Terrasi
}
