% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGuide.R
\name{getGuide}
\alias{getGuide}
\title{Get a guide}
\usage{
getGuide(guideId)
}
\arguments{
\item{guideId}{Id of your guide}
}
\value{
Json
}
\description{
Get a guide
}
\details{
ResCode
401 : invalid key or ip
429 : Too Many Requests
500 : Internal error
550 : Error while generating guide.
250 : Guide generation on the way. Please wait a few minutes.
}
\examples{
\dontrun{
guide <- getGuide(0)
}
}
\author{
Vincent Terrasi
}
