% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetamot.R
\name{getMetamot}
\alias{getMetamot}
\title{List a given metamot batch}
\usage{
getMetamot(metamot_id)
}
\value{
a json with metamot details
}
\description{
List a given metamot batch
}
\details{
List overview of the given metamot batch

ResCode
401 : Unauthorized : user or password does not match.
406 : Request and response must be application/json
}
\examples{
\dontrun{
metamot <- getMetamot(1)
}
}
\author{
Vincent Terrasi
}
