% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSimulatedLexies.R
\name{getSimulatedLexies}
\alias{getSimulatedLexies}
\title{get Simulated Lexies}
\usage{
getSimulatedLexies(df, nbchar)
}
\arguments{
\item{nbchar}{}
}
\value{
a dataframe
}
\description{
get Simulated Lexies
}
\details{

}
\examples{
\dontrun{
lexies <- getSimulatedLexies(df,nbchar)
}
}
\author{
Vincent Terrasi
}
