% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatus.R
\name{getStatus}
\alias{getStatus}
\title{Get account status (number of tokens spent, remaining tokens, current processing)}
\usage{
getStatus()
}
\value{
Json
}
\description{
Get account status (number of tokens spent, remaining tokens, current processing)
}
\details{
ResCode
200 : Ok
401 : invalid key or ip
429 : Too Many Requests
500 : Internal error
}
\examples{
\dontrun{
status <- getStatus()
}
}
\author{
Vincent Terrasi
}
