% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCrawls.R
\name{listCrawl}
\alias{listCrawl}
\title{list Crawls}
\usage{
listCrawl()
}
\value{
Json
}
\description{
list Crawls
}
\details{
ResCode
}
\examples{
\dontrun{
crawls <- listCrawl()
}
}
\author{
Vincent Terrasi
}
