% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listLexies.R
\name{listLexies}
\alias{listLexies}
\title{List the lexies of all expressions of a metamot batch}
\usage{
listLexies(metamot_id)
}
\value{
a json with metamot lexies
}
\description{
List the lexies of all expressions of a metamot batch
}
\details{
ResCode
401 : Unauthorized : user or password does not match.
404 : Unknown uid
406 : Request and response must be application/json
}
\examples{
\dontrun{
status <- listLexies(metamot_id)
}
}
\author{
Vincent Terrasi
}
