% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMetamots.R
\name{listMetamots}
\alias{listMetamots}
\title{list Metamots}
\usage{
listMetamots()
}
\value{
a json with metamot batch Id, language, expressions and timestamp
}
\description{
list Metamots
}
\details{
List all Metamot sets for the user

ResCode
401 : Unauthorized : user or password does not match.
406 : Request and response must be application/json
}
\examples{
\dontrun{
metamots <- listMetamots()
}
}
\author{
Vincent Terrasi
}
