% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreLexies.R
\name{scoreLexies}
\alias{scoreLexies}
\title{scoreLexies}
\usage{
scoreLexies(model, lexie_id, url)
}
\arguments{
\item{model}{Table with all lexies}

\item{lexie_id}{Name of your lexie}

\item{url}{URL of the analysed webpage}
}
\description{
scoreLexies
}
\details{

}
\examples{
\dontrun{

}
}
\author{
Vincent Terrasi
}
