% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unAccent}
\alias{unAccent}
\title{unAccent()}
\usage{
unAccent(text)
}
\value{
a text without accents
}
\description{
unAccent()
}
\details{

}
\examples{
\dontrun{
unAccent(text)
}
}
\author{
Vincent Terrasi
}
