% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{extractText}
\alias{extractText}
\title{extractText}
\usage{
extractText(url, minchar, filter = FALSE)
}
\value{
A character with the main text
}
\description{
extractText
}
\details{

}
\examples{
\dontrun{
extractText("https//www.wikipedia.com",2O,filter=FALSE)
extractText("https//www.wikipedia.com",2O,filter=TRUE)
}
}
\author{
Vincent Terrasi
}
