% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSimulatedLexies.R
\name{listSimulatedLexies}
\alias{listSimulatedLexies}
\title{List the lexies and optimum occurence count for an expression of a metamot batch}
\usage{
listSimulatedLexies(metamot_id, lexie_id)
}
\arguments{
\item{metamot_id}{The batch id}

\item{lexie_id}{The expression to simulate}
}
\value{
a json with lexies occurences for several text sizes.
}
\description{
List the lexies and optimum occurence count for an expression of a metamot batch
}
\details{
ResCode
401 : Unauthorized : user or password does not match.
404 : Unknown uid
406 : Request and response must be application/json
}
\examples{
\dontrun{
lexies <- listSimulatedLexies(metamot_id,"test")
}
}
\author{
Vincent Terrasi
}
