% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unAccent}
\alias{unAccent}
\title{Remove all accents from text}
\usage{
unAccent(text)
}
\arguments{
\item{text}{your text}
}
\value{
a text without accents
}
\description{
Remove all accents from text
}
\details{

}
\examples{
\dontrun{
unAccent(text)
}
}
\author{
Vincent Terrasi
}
